ER-285 :hammer: - Moving EmployerAccount service to the Vacancy Client project as it is to be used from there.
