+++ src/Employer/Employer.Web/Configuration/IoC.cs
@@ -5,8 +5,6 @@ using Esfa.Recruit.Employer.Web.Services;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Options;
-using SFA.DAS.EAS.Account.Api.Client;
 using SFA.DAS.Providers.Api.Client;
 
 namespace Esfa.Recruit.Employer.Web.Configuration
@@ -22,9 +20,7 @@ namespace Esfa.Recruit.Employer.Web.Configuration
             services.Configure<ManageApprenticeshipsRoutes>(configuration.GetSection("ManageApprenticeshipsRoutes"));
             services.AddSingleton<ManageApprenticeshipsLinkHelper>();
             services.Configure<AuthenticationConfiguration>(configuration.GetSection("Authentication"));
-            services.Configure<AccountApiConfiguration>(configuration.GetSection("AccountApiConfiguration"));
 
-            RegisterAccountApiClientDeps(services);
             RegisterProviderApiClientDep(services, configuration);
 
             services.AddSingleton<IHttpContextAccessor, HttpContextAccessor>(); // Used by NLog to log out traceidentifier value.
@@ -36,19 +32,12 @@ namespace Esfa.Recruit.Employer.Web.Configuration
 
         private static void RegisterServiceDeps(IServiceCollection services)
         {
-            services.AddTransient<IEmployerAccountService, EmployerAccountService>();
             services.AddTransient<ITrainingProviderService, TrainingProviderService>();
         }
 
         private static void RegisterProviderApiClientDep(IServiceCollection services, IConfiguration configuration)
         {
             services.AddTransient<IProviderApiClient>(_ => new ProviderApiClient(configuration.GetValue<string>("ProviderApiUrl")));
-        }
-
-        private static void RegisterAccountApiClientDeps(IServiceCollection services)
-        {
-            services.AddSingleton<IAccountApiConfiguration>(kernal => kernal.GetService<IOptions<AccountApiConfiguration>>().Value);
-            services.AddTransient<IAccountApiClient, AccountApiClient>();
         }
 
         private static void RegisterOrchestratorDeps(IServiceCollection services)
