+++ src/Employer/Employer.Web/Employer.Web.csproj
@@ -10,7 +10,6 @@
     <PackageReference Include="NWebsec.AspNetCore.Middleware" Version="1.1.0" />
     <PackageReference Include="NLog.Web.AspNetCore" Version="4.5.0-rc2" />
     <PackageReference Include="NLog" Version="4.5.0-rc04" />
-    <PackageReference Include="SFA.DAS.Account.Api.Client" Version="1.2.0.58438" />
     <PackageReference Include="SFA.DAS.NLog.Targets.Redis" Version="1.0.0.56651" />
     <PackageReference Include="SFA.DAS.Providers.Api.Client" Version="0.10.138" />
     <PackageReference Include="SFA.DAS.Apprenticeships.Api.Types" Version="0.10.139" />
