+++ src/Shared/Recruit.Vacancies.Client/Domain/Entities/EmployerIdentifier.cs
@@ -1,4 +1,4 @@
-﻿namespace Esfa.Recruit.Employer.Web.Models
+﻿namespace Esfa.Recruit.Vacancies.Client.Domain.Entities
 {
     public class EmployerIdentifier
     {
